fixar kopplingar

+++ St-dafint.ab/src/main/java/com/example/nicecleaning/dto/BookingResponseDTO.java
@@ -2,7 +2,7 @@ package com.example.nicecleaning.dto;
 
 public record BookingResponseDTO(
         // ID på städningen
-        Long id,
+        int id,
         // Datum och tid på städningen
         String date,
         String time,
@@ -11,6 +11,6 @@ public record BookingResponseDTO(
         // Statusen på den. Bekräftad, obekräftad, osv.
         int status,
         // Vem som bokade den. Ska bytas ut mot vem städaren är i slutändan
-        Long appUserId) {
+        int appUserId) {
 
 }
